UL-81 - feat(ProfileIndexPage): Finish profile index page
