+++ lib/profile/profile_export.dart
@@ -1,1 +1,1 @@
-export 'profile_index_page.dart';
+export 'profile_index_page/profile_index_page.dart';
